Unified setup.py and setup3k.py, source tree modified (sfml/ -> src/sfml), added a new module x11 (to provide some utilities).
The file setup.py detects which python version is used.
In a source package, source should be in a directory src/. This is now done.
The module X11 has been added to provide utilities that the official X11 wrapper does not. With this module you can flush the screen which is needed when we want to interface with pyqt4.